Cleanup rest modules
