+++ rest-api/src/main/resources/application.conf
@@ -1,6 +1,11 @@
 pipewrench {
+  git {
+    url = "https://github.com/Cargill/pipewrench"
+  }
   directory {
-    output = "pipewrench_config"
-    template = "pipewrench_config/templates"
+    base = "pipewrench_conf"
+    pipewrench = "pipewrench_conf"
+    templates = "pipewrench_conf/pipewrench/templates"
+    ingest = "pipewrench_conf/ingest_conf"
   }
 }
