+++ rest-api/src/main/scala/io/phdata/pipeforge/rest/RestApi.scala
@@ -21,16 +21,13 @@ import io.phdata.pipeforge.rest.module._
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-object RestApi {
+object RestApi extends RestModule {
 
   def start(port: Int): Unit = {
-    val app = new AppModule with ExecutionContextModule with ConfigurationModule with ServiceModule
-    with AkkaModule with HttpModule with RestModule
 
-    import app.executionContext
     Await.ready(
       for {
-        _ <- app.restApi.start(port)
+        _ <- restApi.start(port)
       } yield Unit,
       Duration.Inf
     )
