--- rest-api/src/main/scala/io/phdata/pipeforge/rest/module/ServiceModule.scala
@@ -1,26 +1,0 @@
-/*
- * Copyright 2018 phData Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.phdata.pipeforge.rest.module
-
-import io.phdata.pipeforge.rest.service.{ PipewrenchService, PipewrenchServiceImpl }
-
-trait ServiceModule {
-  this: AkkaModule with ExecutionContextModule with ConfigurationModule with HttpModule =>
-
-  val pipewrenchService: PipewrenchService = new PipewrenchServiceImpl()
-
-}
